refactor collection results
