+++ docs/how-to/custom_views_code.py
@@ -8,7 +8,7 @@ import abc
 from dbally import decorators, MethodsBaseView
 from dbally.audit.event_handlers.cli_event_handler import CLIEventHandler
 from dbally.iql import IQLQuery, syntax
-from dbally.data_models.execution_result import ViewExecutionResult
+from dbally.collections.results import ViewExecutionResult
 from dbally.llms.litellm import LiteLLM
 
 @dataclass
