+++ src/dbally/collections/__init__.py
@@ -1,0 +1,11 @@
+from dbally.collections.base import Collection
+from dbally.collections.exceptions import IndexUpdateError, NoViewFoundError
+from dbally.collections.results import ExecutionResult, ViewExecutionResult
+
+__all__ = [
+    "Collection",
+    "ExecutionResult",
+    "ViewExecutionResult",
+    "NoViewFoundError",
+    "IndexUpdateError",
+]
