+++ src/dbally/collections/results.py
