+++ src/dbally/data_models/audit.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Optional, Union
 
-from dbally.data_models.execution_result import ExecutionResult
+from dbally.collections.results import ExecutionResult
 from dbally.prompts import ChatFormat
