+++ src/dbally/views/freeform/text2sql/view.py
@@ -7,7 +7,7 @@ from typing import Dict, Iterable, List, Optional, Tuple
 from sqlalchemy import ColumnClause, Engine, MetaData, Table, text
 
 from dbally.audit.event_tracker import EventTracker
-from dbally.data_models.execution_result import ViewExecutionResult
+from dbally.collections.results import ViewExecutionResult
 from dbally.llms.base import LLM
 from dbally.llms.clients.base import LLMOptions
 from dbally.prompts import PromptTemplate
